+++ packages/sdk-next/src/api/provider/websocket.ts
@@ -1,7 +1,5 @@
-import { getWsProvider } from 'polkadot-api/ws-provider';
+import { getWsProvider } from 'polkadot-api/ws';
 import { withLogsRecorder } from 'polkadot-api/logs-provider';
-
-import { withLegacy } from '@polkadot-api/legacy-provider';
 
 type WsProviderConfig = Parameters<typeof getWsProvider>[1];
 
@@ -10,10 +8,7 @@ export const getWs = (
   config: WsProviderConfig = {}
 ) => {
   const endpoints = typeof wsUrl === 'string' ? wsUrl.split(',') : wsUrl;
-  const wsProvider = getWsProvider(endpoints, {
-    innerEnhancer: withLegacy(),
-    ...config,
-  });
+  const wsProvider = getWsProvider(endpoints, config);
   withLogsRecorder((line) => console.log(line), wsProvider);
   return wsProvider;
 };