+++ packages/xc-cfg/src/validations/index.ts
@@ -8,8 +8,10 @@ import { FeeValidation, DestFeeValidation } from './base';
 import {
   HubEdValidation,
   HubFrozenValidation,
+  HydrationDepositLimitValidation,
   HydrationEdValidation,
   HydrationMrlFeeValidation,
+  HydrationWithdrawLimitValidation,
 } from './chain';
 
 type Matcher = { [key: string]: TransferValidationConstraint };
@@ -28,4 +30,6 @@ export const validations: TransferValidation[] = [
   new HubFrozenValidation(Matchers.isHub, Matchers.isAny),
   new HydrationEdValidation(Matchers.isAny, Matchers.isHydration),
   new HydrationMrlFeeValidation(Matchers.isHydration, Matchers.isAny),
+  new HydrationDepositLimitValidation(Matchers.isAny, Matchers.isHydration),
+  new HydrationWithdrawLimitValidation(Matchers.isHydration, Matchers.isAny),
 ];