+++ packages/xc/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @galacticcouncil/xc
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [1281cfa]
+  - @galacticcouncil/xc-core@1.0.0
+  - @galacticcouncil/xc-cfg@1.0.0
+  - @galacticcouncil/xc-sdk@1.0.0
+
 ## 0.6.0
 
 ### Minor Changes